@@ -12,31 +12,31 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "axios": "^1.4.0",
-    "mitt": "^3.0.0",
-    "moment": "^2.29.4",
-    "pinia": "^2.1.3",
-    "swiper": "^9.4.1",
+    "axios": "^1.7.2",
+    "mitt": "^3.0.1",
+    "moment": "^2.30.1",
+    "pinia": "^2.1.7",
+    "swiper": "^11.1.4",
     "ts-md5": "^1.3.1",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.2"
+    "vue": "^3.4.27",
+    "vue-router": "^4.3.2"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@tsconfig/node18": "^2.0.1",
-    "@types/node": "^20.3.1",
-    "@vitejs/plugin-vue": "^4.2.3",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "@vue/tsconfig": "^0.4.0",
-    "eslint": "^8.39.0",
-    "eslint-plugin-vue": "^9.11.0",
+    "@rushstack/eslint-patch": "^1.10.3",
+    "@tsconfig/node18": "^18.2.4",
+    "@types/node": "^20.14.0",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/tsconfig": "^0.5.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.26.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.8",
-    "sass": "^1.63.6",
-    "typescript": "~5.1.3",
-    "vite": "^4.3.9",
-    "vite-plugin-pwa": "^0.16.4",
-    "vue-tsc": "^1.6.5"
+    "prettier": "^3.3.0",
+    "sass": "^1.77.4",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.12",
+    "vite-plugin-pwa": "^0.20.0",
+    "vue-tsc": "^2.0.19"
   }
 }
