@@ -1,8 +1,8 @@
 import {fetch} from '../utils/http';
 import  type { IData, IAd } from './model';
 
-export const getAds = (position: number) => fetch<IData<IAd>>('shop/ad', {
+export const getAds = (position: number) => fetch<IData<IAd>>('ad', {
     position,
 });
 
-export const getBanners = () => fetch<IData<IAd>>('shop/ad/banner');
+export const getBanners = () => fetch<IData<IAd>>('ad/banner');
