@@ -22,9 +22,9 @@
         <div class="dialog-bg" v-if="calendarVisible" @click="hideCalerdar"/>
     </div>
 </template>
-<script lang="ts">
-import { each } from '@/utils';
-import { defineComponent } from 'vue'
+<script setup lang="ts">
+import { eachObject } from '@/utils';
+import { ref, watch } from 'vue'
 
 interface IColumn {
     index: number,
@@ -37,276 +37,278 @@ interface IPoint {
     y: number,
 }
 
-export default defineComponent({
-    props: {
-        column: {type: Number, default: 1},
-        title: {type: String, default: '请选择'},
-        textTag: {type: String, default: 'name'},
-        valueTag: {type: String, default: 'id'},
-        childrenTag: {type: String, default: 'children'},
-        lineHeight: {type: Number, default: 30},
-        value: {type: [Array, Object]},
-        items: {type: Object, required: true},
-    },
-    data() {
-        const columnList: IColumn[] = [];
-        const startPoint: IPoint = {x: 0, y: 0};
-        return {
-            columnList,
-            calendarVisible: false,
-            startPoint
-        }
-    },
-    watch: {
-        items(val: any, oldVal: any) {
-            if (Object.keys(val).length === 0) {
-                return;
-            }
-            if (Object.keys(oldVal).length === 0 && this.calendarVisible) {
-                this.refresh();
-            }
-        }
-    },
-    methods: {
-        showCalendar() {
-            this.calendarVisible = true;
-            this.refresh();
-        },
-        hideCalerdar() {
-            this.calendarVisible = false;
-        },
-        tapItem(index: number, i: number) {
-            this.columnList[index].index = i;
-            this.columnList[index].style = this.getIndexStyle(i);
-            for (let j = index + 1; j < this.columnList.length; j ++) {
-                this.refreshColumn(j, 0);
-            }
-        },
-        touchStart(event: TouchEvent) {
-            const touch = event.targetTouches[0];
-            this.startPoint = {
-                x: touch.pageX,
-                y: touch.pageY,
-            };
-        },
-        touchMove(event: TouchEvent) {
-            const touch = event.targetTouches[0];
-            if (event.targetTouches.length > 1 || ((event as any).scale && (event as any).scale !== 1)) {
-                return;
-            }
-            event.preventDefault();
-            const y = touch.pageY - this.startPoint.y;
-            const diff = Math.abs(y);
-            if (diff >= this.lineHeight) {
-                // 滑动了一个单位就更新起始y 坐标
-                this.startPoint.y = touch.pageY;
-                this.doMove(diff, y < 0, this.startPoint.x);
-            }
-        },
-        getClientWidth(): number {
-            if (window.innerHeight !== undefined) {
-                return window.innerWidth;
-            }
-            if (document.compatMode === 'CSS1Compat') {
-                return document.documentElement.clientWidth;
-            }
-            return document.body.clientWidth;
-        },
-        /**
-        * 滑动
-        * @param distance 距离的绝对值
-        * @param isUp 是否是上滑
-        * @param x 触发的位置，自动定位到第几级
-        */
-        doMove(distance: number, isUp = true, x = 0) {
-            const diff: number = isUp ? Math.floor(distance / this.lineHeight) : - Math.ceil(distance / this.lineHeight);
-            let column = 0;
-            if (diff === 0) {
-                return this;
-            }
-            if (this.column > 1) {
-                column = Math.floor(x / (this.getClientWidth() / this.column));
-            }
-            this.tapItem(column, this.columnList[column].index + diff);
-        },
-        tapOutput() {
-            this.hideCalerdar();
-            if (this.columnList.length < 1) {
-                return;
-            }
-            const data = [];
-            const name = [];
-            for (const column of this.columnList) {
-                const current = column.index || 0;
-                if (!column.items || column.items.length < current) {
-                    break;
-                }
-                name.push(column.items[current][this.textTag]);
-                data.push(this.cloneItem(column.items[current]));
-            }
-            if (typeof this.value === 'object' && this.value instanceof Array) {
-                this.$emit('input', data);
-                return;
-            }
-            if (data.length < 1) {
-                return;
-            }
-            const item: any = data[data.length - 1];
-            if (!Object.prototype.hasOwnProperty.call(item, 'full_name')) {
-                item.full_name = name.join(' ');
-            }
-            this.$emit('input', item);
-        },
-        cloneItem(obj: any): any {
-            const item: any = new Object();
-            for (const key in obj) {
-                if (Object.prototype.hasOwnProperty.call(obj, key) && key !== this.childrenTag) {
-                    item[key] = obj[key];
-                }
-            }
-            return item;
-        },
-        refresh() {
-            if (Object.keys(this.items).length === 0) {
-                return;
-            }
-            const path = this.getPath();
-            for (let i = 0; i < this.column; i++) {
-                this.refreshColumn(i, path.length > i ? path[i] : 0);
-            }
-        },
-        refreshColumn(index: number, selected = 0) {
-            const data: any[] = [];
-            each(this.getColumnData(index), (item) => {
-                data.push(item);
-            });
-            if (this.columnList.length <= index) {
-                this.columnList.push({
-                    index: selected,
-                    items: data,
-                    style: this.getIndexStyle(selected),
-                });
-                return;
-            }
-            this.columnList[index].items = data;
-            this.columnList[index].index = selected;
-            this.columnList[index].style = this.getIndexStyle(selected);
-        },
-        getColumnData(index: number) {
-            if (index < 1) {
-                return this.items;
-            }
-            if (this.columnList.length < index) {
-                return [];
-            }
-            const column = this.columnList[index - 1];
-            const current = column.index || 0;
-            if (!column.items || column.items.length <= current) {
-                return [];
-            }
-            const item = column.items[current];
-            if (!Object.prototype.hasOwnProperty.call(item, this.childrenTag)) {
-                return [];
-            }
-            return item[this.childrenTag];
-        },
-        getIndexStyle(index: number) {
-            const top = 2 * this.lineHeight - index  * this.lineHeight;
-            return 'transform: translate(0px, ' + top + 'px) translateZ(0px)';
-        },
-         /**
-        * 根据ID查找无限树的路径
-        * @param id
-        */
-        getPath(): number[] {
-            if (!this.value) {
-                return [];
-            }
-            let id: string | number;
-            if (typeof this.value === 'object') {
-                if (this.value instanceof Array) {
-                    return this.getPathByArr();
-                }
-                id = (this.value as any)[this.valueTag];
-            } else {
-                id = this.value;
-            }
-            if (!id) {
-                return [];
-            }
-            const path: number[] = [];
-            let found = false;
-            const findPath = (data: any) => {
-                if (typeof data !== 'object') {
-                    return;
-                }
-                let iii = -1;
-                each(data, (args, key) => {
-                    iii ++;
-                    if (key === id || args[this.valueTag] === id) {
-                        path.push(iii);
-                        found = true;
-                        return false;
-                    }
-                    if (!Object.prototype.hasOwnProperty.call(args, this.childrenTag)) {
-                        return;
-                    }
-                    findPath(args[this.childrenTag]);
-                    if (found) {
-                        path.push(iii);
-                        return false;
-                    }
+const emit = defineEmits(['update:modalValue'])
+const props = withDefaults(defineProps<{
+    column?: number,
+    title?: string,
+    textTag?: string,
+    valueTag?: string,
+    childrenTag?: string,
+    lineHeight?: number,
+    modalValue?: any,
+    items: any,
+}>(), {
+    column: 1,
+    title: '请选择',
+    textTag: 'name',
+    valueTag: 'id',
+    childrenTag: 'children',
+    lineHeight: 30,
+});
 
-                });
-            };
-            let ii = -1;
+const columnList = ref<IColumn[]>([]);
+const startPoint = ref<IPoint>({x: 0, y: 0});
+const calendarVisible = ref(false);
 
-            each(this.items, (data: any) => {
-                ii ++;
-                findPath(data[this.childrenTag]);
-                if (found) {
-                    path.push(ii);
-                    return false;
-                }
-            });
-            path.reverse();
-            return path;
-        },
-        getPathByArr() {
-            let data = this.items;
-            const path: number[] = [];
-            each(this.value, (item) => {
-                const [index, children] = this.getIndexWithItems(item, data);
-                path.push(index);
-                if (!children) {
-                    return false;
-                }
-                data = children;
-            });
-            return path;
-        },
-        getIndexWithItems(item: any, data: any) {
-            if (!data) {
-                return [0, undefined];
-            }
-            let i = -1;
-            let index = -1;
-            let children: any;
-            each(data, (obj) => {
-                i ++;
-                if (!item[this.valueTag] && item[this.valueTag] === obj[this.valueTag]) {
-                    children = obj[this.childrenTag];
-                    index = i;
-                    return false;
-                }
-            });
-            if (index < 0) {
-                return [0, undefined];
-            }
-            return [index, children];
+function showCalendar() {
+    calendarVisible.value = true;
+    refresh();
+}
+function hideCalerdar() {
+    calendarVisible.value = false;
+}
+function tapItem(index: number, i: number) {
+    columnList.value[index].index = i;
+    columnList.value[index].style = getIndexStyle(i);
+    for (let j = index + 1; j < columnList.value.length; j ++) {
+        refreshColumn(j, 0);
+    }
+}
+function touchStart(event: TouchEvent) {
+    const touch = event.targetTouches[0];
+    startPoint.value = {
+        x: touch.pageX,
+        y: touch.pageY,
+    };
+}
+function touchMove(event: TouchEvent) {
+    const touch = event.targetTouches[0];
+    if (event.targetTouches.length > 1 || ((event as any).scale && (event as any).scale !== 1)) {
+        return;
+    }
+    event.preventDefault();
+    const y = touch.pageY - startPoint.value.y;
+    const diff = Math.abs(y);
+    if (diff >= props.lineHeight) {
+        // 滑动了一个单位就更新起始y 坐标
+        startPoint.value.y = touch.pageY;
+        doMove(diff, y < 0, startPoint.value.x);
+    }
+}
+function getClientWidth(): number {
+    if (window.innerHeight !== undefined) {
+        return window.innerWidth;
+    }
+    if (document.compatMode === 'CSS1Compat') {
+        return document.documentElement.clientWidth;
+    }
+    return document.body.clientWidth;
+}
+/**
+* 滑动
+* @param distance 距离的绝对值
+* @param isUp 是否是上滑
+* @param x 触发的位置，自动定位到第几级
+*/
+function doMove(distance: number, isUp = true, x = 0) {
+    const diff: number = isUp ? Math.floor(distance / props.lineHeight) : - Math.ceil(distance / props.lineHeight);
+    let column = 0;
+    if (diff === 0) {
+        return;
+    }
+    if (column > 1) {
+        column = Math.floor(x / (getClientWidth() / column));
+    }
+    tapItem(column, columnList.value[column].index + diff);
+}
+function tapOutput() {
+    hideCalerdar();
+    if (columnList.value.length < 1) {
+        return;
+    }
+    const data = [];
+    const name = [];
+    for (const column of columnList.value) {
+        const current = column.index || 0;
+        if (!column.items || column.items.length < current) {
+            break;
+        }
+        name.push(column.items[current][props.textTag]);
+        data.push(cloneItem(column.items[current]));
+    }
+    if (typeof props.modalValue === 'object' && props.modalValue instanceof Array) {
+        emit('update:modalValue', data);
+        return;
+    }
+    if (data.length < 1) {
+        return;
+    }
+    const item: any = data[data.length - 1];
+    if (!Object.prototype.hasOwnProperty.call(item, 'full_name')) {
+        item.full_name = name.join(' ');
+    }
+    emit('update:modalValue', item);
+}
+function cloneItem(obj: any): any {
+    const item: any = new Object();
+    for (const key in obj) {
+        if (Object.prototype.hasOwnProperty.call(obj, key) && key !== props.childrenTag) {
+            item[key] = obj[key];
         }
     }
+    return item;
+}
+function refresh() {
+    if (Object.keys(props.items).length === 0) {
+        return;
+    }
+    const path = getPath();
+    for (let i = 0; i < props.column; i++) {
+        refreshColumn(i, path.length > i ? path[i] : 0);
+    }
+}
+function refreshColumn(index: number, selected = 0) {
+    const data: any[] = [];
+    eachObject(getColumnData(index), (item) => {
+        data.push(item);
+    });
+    if (columnList.value.length <= index) {
+        columnList.value.push({
+            index: selected,
+            items: data,
+            style: getIndexStyle(selected),
+        });
+        return;
+    }
+    columnList.value[index].items = data;
+    columnList.value[index].index = selected;
+    columnList.value[index].style = getIndexStyle(selected);
+}
+function getColumnData(index: number) {
+    if (index < 1) {
+        return props.items;
+    }
+    if (columnList.value.length < index) {
+        return [];
+    }
+    const column = columnList.value[index - 1];
+    const current = column.index || 0;
+    if (!column.items || column.items.length <= current) {
+        return [];
+    }
+    const item = column.items[current];
+    if (!Object.prototype.hasOwnProperty.call(item, props.childrenTag)) {
+        return [];
+    }
+    return item[props.childrenTag];
+}
+function getIndexStyle(index: number) {
+    const top = 2 * props.lineHeight - index  * props.lineHeight;
+    return 'transform: translate(0px, ' + top + 'px) translateZ(0px)';
+}
+    /**
+* 根据ID查找无限树的路径
+* @param id
+*/
+function getPath(): number[] {
+    if (!props.modalValue) {
+        return [];
+    }
+    let id: string | number;
+    if (typeof props.modalValue === 'object') {
+        if (props.modalValue instanceof Array) {
+            return getPathByArr();
+        }
+        id = (props.modalValue as any)[props.valueTag];
+    } else {
+        id = props.modalValue;
+    }
+    if (!id) {
+        return [];
+    }
+    const path: number[] = [];
+    let found = false;
+    const findPath = (data: any) => {
+        if (typeof data !== 'object') {
+            return;
+        }
+        let iii = -1;
+        eachObject(data, (args, key) => {
+            iii ++;
+            if (key === id || args[props.valueTag] === id) {
+                path.push(iii);
+                found = true;
+                return false;
+            }
+            if (!Object.prototype.hasOwnProperty.call(args, props.childrenTag)) {
+                return;
+            }
+            findPath(args[props.childrenTag]);
+            if (found) {
+                path.push(iii);
+                return false;
+            }
+
+        });
+    };
+    let ii = -1;
+
+    eachObject(props.items, (data: any) => {
+        ii ++;
+        findPath(data[props.childrenTag]);
+        if (found) {
+            path.push(ii);
+            return false;
+        }
+    });
+    path.reverse();
+    return path;
+}
+function getPathByArr() {
+    let data = props.items;
+    const path: number[] = [];
+    eachObject(props.modalValue, (item) => {
+        const [index, children] = getIndexWithItems(item, data);
+        path.push(index);
+        if (!children) {
+            return false;
+        }
+        data = children;
+    });
+    return path;
+}
+function getIndexWithItems(item: any, data: any) {
+    if (!data) {
+        return [0, undefined];
+    }
+    let i = -1;
+    let index = -1;
+    let children: any;
+    eachObject(data, (obj) => {
+        i ++;
+        if (!item[props.valueTag] && item[props.valueTag] === obj[props.valueTag]) {
+            children = obj[props.childrenTag];
+            index = i;
+            return false;
+        }
+    });
+    if (index < 0) {
+        return [0, undefined];
+    }
+    return [index, children];
+}
+
+
+watch(() => props.items, (val: any, oldVal: any) => {
+    if (Object.keys(val).length === 0) {
+        return;
+    }
+    if (Object.keys(oldVal).length === 0 && calendarVisible.value) {
+        refresh();
+    }
 });
+
+
 </script>
 <style lang="scss" scoped>
 .dialog-select-input {
