@@ -1,5 +1,6 @@
-import './assets/css/theme.scss'
 import './assets/iconfont/iconfont.css';
+import './assets/css/style.scss'
+
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
@@ -8,6 +9,8 @@ import App from './App.vue'
 import router from './router'
 import emitter from './event';
 import { createDialog } from './components/Dialog/plugin'
+import { createDirective } from './directives';
+import { createSerive } from './services';
 
 
 const app = createApp(App, {
@@ -18,6 +21,8 @@ const app = createApp(App, {
 
 app.use(createPinia())
 app.use(createDialog())
+app.use(createDirective())
+app.use(createSerive())
 app.use(router)
 
 app.mount('#app')
