@@ -1,13 +1,5 @@
-import { assetUri } from '../utils'
+import { useAuth } from "../services";
+
 export function assetsFilter(value: string): string {
-    if (!value) {
-        return '';
-    }
-    if (value.indexOf('//') >= 0) {
-        return value;
-    }
-    if (value.startsWith('/')) {
-        return assetUri + value;
-    }
-    return assetUri + '/' + value;
+    return useAuth().assetUri(value) ?? '';
 }
